# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "API for social networking platform",
    "title": "Social networking platform"
  },
  "servers": [
    {
      "url": "http://localhost:3500"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "cookie",
        "name": "jwtToken",
        "type": "apiKey"
      }
    }
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get User Posts",
        "description": "User will get posts posted by them",
        "security": [{ "JWT": {} }],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Add Posts",
        "description": "User can add new posts",
        "security": [{ "JWT": {} }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/bookmarks": {
      "get": {
        "tags": ["Bookmarks"],
        "summary": "Retrieve all bookmarked posts",
        "description": "User can retrieve all of his/her bookmarked posts",
        "security": [{ "JWT": {} }],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "No bookmarked posts found"
          }
        }
      }
    },
...
  }
}